Split up NodeTransformer into many classes
