+++ src/argon/virtualization/virtualizer/virtualizer_ifthenelse.py
@@ -1,12 +1,9 @@
 import ast
-from collections import namedtuple
 import dis
-import types
 import typing
 
 from argon.block import Block
-from argon.node.control import IfThenElse, Loop
-from argon.node.function_call import FunctionCall
+from argon.node.control import IfThenElse
 from argon.node.phi import Phi
 from argon.node.undefined import Undefined
 from argon.ref import Exp, Ref
@@ -14,7 +11,7 @@ from argon.srcctx import SrcCtx
 from argon.state import ScopeContext, State, stage
 from argon.types.boolean import Boolean
 from argon.types.null import Null
-from argon.virtualization.type_mapper import concrete_to_abstract
+from argon.virtualization.virtualizer.virtualizer_base import TransformerBase
 
 
 def stage_undefined(name, T, file_name, lineno, col_offset):
@@ -98,273 +95,7 @@ def stage_if(
     )
 
 
-def stage_function_call(
-    func: types.FunctionType, args: typing.List[typing.Any]
-) -> Ref[typing.Any, typing.Any]:
-    white_list = [
-        print
-    ]  # TODO: Add more whitelisted functions here that we don't want to stage
-    if func in white_list:
-        return func(*args)
-
-    abstract_args = [concrete_to_abstract(arg) for arg in args]
-    abstract_func = concrete_to_abstract.function(func, abstract_args)
-    return stage(
-        FunctionCall[abstract_func.F](abstract_func, abstract_args), ctx=SrcCtx.new(2)
-    )
-
-
-def stage_loop(
-    file_name: str,
-    lineno: int,
-    col_offset: int,
-    values: typing.List[Exp[typing.Any, typing.Any]],
-    binds: typing.List[Exp[typing.Any, typing.Any]],
-    cond_scope_context: ScopeContext,
-    cond: Exp[typing.Any, typing.Any],
-    loop_scope_context: ScopeContext,
-    outputs: namedtuple,
-) -> Ref[typing.Any, typing.Any]:
-    condBlk = Block[Boolean](
-        cond_scope_context.scope.inputs, cond_scope_context.scope.symbols, cond
-    )
-    bodyBlk = Block[Null](
-        loop_scope_context.scope.inputs,
-        loop_scope_context.scope.symbols,
-        Null().const(None),
-    )
-    return stage(
-        Loop(values, binds, condBlk, bodyBlk, outputs),
-        ctx=SrcCtx(file_name, dis.Positions(lineno=lineno, col_offset=col_offset)),
-    )
-
-
-class VariableTrackerWhile:
-    def __init__(self, variable_tracker: "VariableTracker"):
-        self.variable_tracker = variable_tracker
-
-    def __enter__(self):
-        self.variable_tracker.push_context()
-        self.variable_tracker.binds_stack.append(set())
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        popped_write_set, _ = self.variable_tracker.pop_context()
-        popped_binds = self.variable_tracker.binds_stack.pop()
-        for var in popped_write_set:
-            if var in popped_binds:
-                self.variable_tracker.add_read_var(var)
-            self.variable_tracker.add_written_var(var)
-
-    def add_written_var(self, var):
-        if (
-            (
-                len(self.variable_tracker.write_set_stack) > 1
-                and len(self.variable_tracker.read_set_stack) > 1
-            )
-            and (
-                var not in self.variable_tracker.current_write_set()
-                and var not in self.variable_tracker.previous_write_set()
-            )
-            and (
-                var in self.variable_tracker.current_read_set()
-                or var in self.variable_tracker.previous_read_set()
-            )
-        ):
-            self.variable_tracker.current_binds().add(var)
-
-
-class VariableTracker:
-    def __init__(self):
-        self.write_set_stack = [set()]
-        self.read_set_stack = [set()]
-        self.binds_stack = []
-        self.variable_tracker_while = VariableTrackerWhile(self)
-
-    def push_context(self):
-        self.write_set_stack.append(set())
-        self.read_set_stack.append(set())
-
-    def pop_context(self):
-        return self.write_set_stack.pop(), self.read_set_stack.pop()
-
-    def fold_context(self):
-        curr_write_set = self.write_set_stack.pop()
-        self.current_write_set().update(curr_write_set)
-        curr_read_set = self.read_set_stack.pop()
-        self.current_read_set().update(curr_read_set)
-
-    def current_write_set(self):
-        return self.write_set_stack[-1]
-
-    def previous_write_set(self):
-        return self.write_set_stack[-2]
-
-    def current_read_set(self):
-        return self.read_set_stack[-1]
-
-    def previous_read_set(self):
-        return self.read_set_stack[-2]
-
-    def current_binds(self):
-        return self.binds_stack[-1]
-
-    def add_written_var(self, var):
-        self.variable_tracker_while.add_written_var(var)
-        self.current_write_set().add(var)
-
-    def add_read_var(self, var):
-        self.current_read_set().add(var)
-
-
-class Transformer(ast.NodeTransformer):
-    def __init__(self, file_name, calls, ifs, if_exps, loops):
-        super().__init__()
-        self.counter = 0
-        self.unique_prefix = "__________"
-        self.file_name = file_name
-        self.calls = calls
-        self.ifs = ifs
-        self.if_exps = if_exps
-        self.loops = loops
-        self.concrete_to_abstract_flag = False
-        self.variable_tracker = VariableTracker()
-
-    def concrete_to_abstract(self, node):
-        return ast.Call(
-            func=ast.Attribute(
-                value=ast.Attribute(
-                    value=ast.Attribute(
-                        value=ast.Attribute(
-                            value=ast.Name(id="__________argon", ctx=ast.Load()),
-                            attr="argon",
-                            ctx=ast.Load(),
-                        ),
-                        attr="virtualization",
-                        ctx=ast.Load(),
-                    ),
-                    attr="type_mapper",
-                    ctx=ast.Load(),
-                ),
-                attr="concrete_to_abstract",
-                ctx=ast.Load(),
-            ),
-            args=[node],
-            keywords=[],
-        )
-
-    def visit_Constant(self, node):
-        if self.concrete_to_abstract_flag:
-            return self.concrete_to_abstract(node)
-        return node
-
-    def visit_Name(self, node):
-        # Save the loaded variables
-        if isinstance(node.ctx, ast.Load):
-            self.variable_tracker.add_read_var(node.id)
-
-        if self.concrete_to_abstract_flag:
-            return self.concrete_to_abstract(node)
-        return node
-
-    def visit_Assign(self, node):
-        # Visit RHS of assignment
-        node.value = self.visit(node.value)
-
-        # Recursively process each target to extract all written variables
-        for target in node.targets:
-            self._process_target(target)
-
-        return node
-
-    def visit_AugAssign(self, node):
-        # Visit RHS of assignment
-        node.value = self.visit(node.value)
-
-        if isinstance(node.target, ast.Name):
-            self.variable_tracker.add_written_var(node.target.id)
-        else:
-            raise NotImplementedError(
-                "Only support single-variable assignments for augmented assignments"
-            )
-
-        return node
-
-    def _process_target(self, target):
-        # Handle simple variable assignment
-        if isinstance(target, ast.Name):
-            self.variable_tracker.add_written_var(target.id)
-        # Handle tuple or list unpacking
-        elif isinstance(target, (ast.Tuple, ast.List)):
-            for element in target.elts:
-                self._process_target(element)
-        else:
-            raise NotImplementedError("Unsupported target type on LHS of assignment")
-
-    def visit_NamedExpr(self, node):
-        # Visit RHS of assignment
-        node.value = self.visit(node.value)
-
-        if isinstance(node.target, ast.Name):
-            self.variable_tracker.add_written_var(node.target.id)
-        else:
-            raise NotImplementedError(
-                "Only support single-variable assignments for walrus operators"
-            )
-
-        return node
-
-    def visit_Break(self, node):
-        raise NotImplementedError("Does not support break statements")
-
-    def visit_Continue(self, node):
-        raise NotImplementedError("Does not support continue statements")
-
-    # This method is called for function calls
-    def visit_Call(self, node):
-        # Recursively visit arguments
-        prev_concrete_to_abstract_flag = self.concrete_to_abstract_flag
-        self.concrete_to_abstract_flag = False
-        self.generic_visit(node)
-        self.concrete_to_abstract_flag = prev_concrete_to_abstract_flag
-
-        # Do not stage the function call if the flag is set to False
-        if not self.calls:
-            if not self.concrete_to_abstract_flag:
-                return node
-            else:
-                return self.concrete_to_abstract(node)
-
-        # Wrap arguments in a list
-        args_list = ast.List(elts=node.args, ctx=ast.Load())
-
-        # Create the staged function call
-        staged_call = ast.Call(
-            func=ast.Attribute(
-                value=ast.Attribute(
-                    value=ast.Attribute(
-                        value=ast.Attribute(
-                            value=ast.Name(id="__________argon", ctx=ast.Load()),
-                            attr="argon",
-                            ctx=ast.Load(),
-                        ),
-                        attr="virtualization",
-                        ctx=ast.Load(),
-                    ),
-                    attr="virtualizer",
-                    ctx=ast.Load(),
-                ),
-                attr="stage_function_call",
-                ctx=ast.Load(),
-            ),
-            args=[node.func, args_list],
-            keywords=[],
-        )
-
-        # Attach source location (line numbers and column offsets)
-        ast.copy_location(staged_call, node)
-
-        return staged_call
-
+class TransformerIfThenElse(TransformerBase):
     # This method is called for ternary "if" expressions
     def visit_IfExp(self, node):
         # Recursively visit the condition, the then case, and the else case
@@ -386,14 +117,18 @@ class Transformer(ast.NodeTransformer):
                 value=ast.Attribute(
                     value=ast.Attribute(
                         value=ast.Attribute(
-                            value=ast.Name(id="__________argon", ctx=ast.Load()),
-                            attr="argon",
+                            value=ast.Attribute(
+                                value=ast.Name(id="__________argon", ctx=ast.Load()),
+                                attr="argon",
+                                ctx=ast.Load(),
+                            ),
+                            attr="virtualization",
                             ctx=ast.Load(),
                         ),
-                        attr="virtualization",
+                        attr="virtualizer",
                         ctx=ast.Load(),
                     ),
-                    attr="virtualizer",
+                    attr="virtualizer_ifthenelse",
                     ctx=ast.Load(),
                 ),
                 attr="stage_if_exp_with_scopes",
@@ -507,7 +242,7 @@ try:
     {temp_var} = lambda _, {var} = {var}: __________argon.argon.virtualization.type_mapper.concrete_to_abstract({var})
     {temp_var_exists} = True
 except NameError:
-    {temp_var} = lambda T : __________argon.argon.virtualization.virtualizer.stage_undefined('{var}', T, '{self.file_name}', {node.lineno}, {node.col_offset})
+    {temp_var} = lambda T : __________argon.argon.virtualization.virtualizer.virtualizer_ifthenelse.stage_undefined('{var}', T, '{self.file_name}', {node.lineno}, {node.col_offset})
     {temp_var_exists} = False
 """
                 ).body
@@ -665,7 +400,7 @@ except NameError:
         # Stage if call
         new_body.extend(
             ast.parse(
-                f"__________argon.argon.virtualization.virtualizer.stage_if('{self.file_name}', {node.lineno}, {node.col_offset}, {cond_scope_name}, {cond_name}, {then_scope_name}, {else_scope_name})"
+                f"__________argon.argon.virtualization.virtualizer.virtualizer_ifthenelse.stage_if('{self.file_name}', {node.lineno}, {node.col_offset}, {cond_scope_name}, {cond_name}, {then_scope_name}, {else_scope_name})"
             ).body
         )
 
@@ -675,7 +410,7 @@ except NameError:
             temp_var_else = self.generate_temp_var(var, "else")
             new_body.extend(
                 ast.parse(
-                    f"{var} = __________argon.argon.virtualization.virtualizer.stage_phi({cond_name}, {temp_var_then}, {temp_var_else})"
+                    f"{var} = __________argon.argon.virtualization.virtualizer.virtualizer_ifthenelse.stage_phi({cond_name}, {temp_var_then}, {temp_var_else})"
                 ).body
             )
 
@@ -724,183 +459,7 @@ except NameError:
         self.variable_tracker.fold_context()
 
         return node
-
-    def visit_While(self, node):
-        with self.variable_tracker.variable_tracker_while:
-            # Increment counter to ensure unique names for this while loop
-            self.counter += 1
-
-            # Properly set flags before recursive visits
-            prev_concrete_to_abstract_flag = self.concrete_to_abstract_flag
-            self.concrete_to_abstract_flag = self.loops
-
-            # Recursively visit the condition
-            self.variable_tracker.push_context()
-            node.test = self.visit(node.test)
-            cond_write_set = self.variable_tracker.current_write_set()
-            cond_read_set = self.variable_tracker.current_read_set()
-            self.variable_tracker.fold_context()
-
-            # Recursively visit the body
-            self.variable_tracker.push_context()
-            node.body = [self.visit(stmt) for stmt in node.body]
-            body_write_set = self.variable_tracker.current_write_set()
-            body_read_set = self.variable_tracker.current_read_set()
-            self.variable_tracker.fold_context()
-
-            if node.orelse:
-                raise NotImplementedError("Does not support else statements for loops")
-
-            new_body: typing.List[ast.stmt] = []
-
-            # Create temporary list of input values and binds
-            values_name = self.generate_temp_var("values")
-            binds_name = self.generate_temp_var("binds")
-            new_body.append(
-                ast.Assign(
-                    targets=[ast.Name(id=values_name, ctx=ast.Store())],
-                    value=ast.List(elts=[], ctx=ast.Load()),
-                )
-            )
-            new_body.append(
-                ast.Assign(
-                    targets=[ast.Name(id=binds_name, ctx=ast.Store())],
-                    value=ast.List(elts=[], ctx=ast.Load()),
-                )
-            )
-
-            # Create binds
-            # TODO: This part needs to be changed to only create binds for inputs
-            for var in self.variable_tracker.current_binds():
-                new_body.extend(
-                    ast.parse(
-                        f"""
-try:
-    {values_name}.append(__________argon.argon.virtualization.type_mapper.concrete_to_abstract({var}))
-    {var} = __________argon.argon.virtualization.type_mapper.concrete_to_abstract({var}).bound('{var}')
-    {binds_name}.append({var})
-except NameError:
-    pass
-"""
-                    ).body
-                )
-
-            # Run the condition under a different scope
-            cond_scope_name = self.generate_temp_var("cond", "scope")
-            cond_name = self.generate_temp_var("cond")
-            new_body.extend(
-                ast.parse(
-                    f"{cond_scope_name} = __________argon.argon.state.State.get_current_state().new_scope()"
-                ).body
-            )
-            new_body.append(
-                ast.With(
-                    items=[
-                        ast.withitem(
-                            context_expr=ast.Name(id=cond_scope_name, ctx=ast.Load())
-                        )
-                    ],
-                    body=[
-                        ast.Assign(
-                            targets=[ast.Name(id=cond_name, ctx=ast.Store())],
-                            value=node.test,
-                        )
-                    ],
-                )
-            )
-
-            # Create a new scope to run the loop body
-            loop_scope_name = self.generate_temp_var("loop", "scope")
-            new_body.extend(
-                ast.parse(
-                    f"{loop_scope_name} = __________argon.argon.state.State.get_current_state().new_scope()"
-                ).body
-            )
-            loop_outputs_name = self.generate_temp_var("loop", "outputs")
-            loop_body = node.body.copy()
-            loop_body.append(  # The loop body should be the original loop body + the following line to assign the loop outputs
-                ast.Assign(
-                    targets=[ast.Name(id=loop_outputs_name, ctx=ast.Store())],
-                    value=ast.Call(
-                        func=ast.Call(
-                            func=ast.Name(id="namedtuple", ctx=ast.Load()),
-                            args=[
-                                ast.Constant(value=loop_outputs_name),
-                                ast.List(
-                                    elts=[
-                                        ast.Constant(value=var)
-                                        for var in self.variable_tracker.current_write_set()
-                                    ],
-                                    ctx=ast.Load(),
-                                ),
-                            ],
-                            keywords=[],
-                        ),
-                        args=[
-                            ast.Name(id=var, ctx=ast.Load())
-                            for var in self.variable_tracker.current_write_set()
-                        ],
-                        keywords=[],
-                    ),
-                )
-            )
-            new_body.append(
-                ast.With(
-                    items=[
-                        ast.withitem(
-                            context_expr=ast.Name(id=loop_scope_name, ctx=ast.Load())
-                        )
-                    ],
-                    body=loop_body,
-                )
-            )
-
-            # Stage the loop
-            new_body.extend(
-                ast.parse(
-                    f"__________argon.argon.virtualization.virtualizer.stage_loop('{self.file_name}', {node.lineno}, {node.col_offset}, {values_name}, {binds_name}, {cond_scope_name}, {cond_name}, {loop_scope_name}, {loop_outputs_name})"
-                ).body
-            )
-
-            # Delete all temporary variables
-            # TODO: COMPLETE THIS PART
-            new_body.append(
-                ast.Delete(targets=[ast.Name(id=values_name, ctx=ast.Del())])
-            )
-            new_body.append(
-                ast.Delete(targets=[ast.Name(id=binds_name, ctx=ast.Del())])
-            )
-            new_body.append(
-                ast.Delete(targets=[ast.Name(id=cond_scope_name, ctx=ast.Del())])
-            )
-            new_body.append(ast.Delete(targets=[ast.Name(id=cond_name, ctx=ast.Del())]))
-            new_body.append(
-                ast.Delete(targets=[ast.Name(id=loop_scope_name, ctx=ast.Del())])
-            )
-            new_body.append(
-                ast.Delete(targets=[ast.Name(id=loop_outputs_name, ctx=ast.Del())])
-            )
-
-            new_body.append(ast.Break())
-
-            # Do not stage the while loop if the flag is set to False
-            # TODO: error handling if condition is Argon type
-            if self.loops:
-                # Replace the original while body with the new wrapped body
-                node.body = new_body
-
-                # Set the condition to True and add a break statement at the end of the body
-                # This lets us run the body just once
-                node.test = ast.Constant(value=True)
-
-            self.counter -= 1
-            self.concrete_to_abstract_flag = prev_concrete_to_abstract_flag
-
-        return node
-
-    def generate_temp_var(self, *args) -> str:
-        return self.unique_prefix + "_".join(args) + "_" + str(self.counter)
-
+    
     def modify_body(
         self, body: typing.List[ast.stmt], scope_name: str, vars: set
     ) -> list:
